Little refactor and deleted settings page
