+++ lib/screens/homepage/widget/FilterPremium.dart
@@ -1,9 +1,9 @@
-import 'package:my_camping/models/popularFilterList.dart';
+import 'package:my_camping/models/filtersList.dart';
 
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import '../../../utilities/appTheme.dart';
-import '../../../models/popularFilterList.dart';
+import '../../../models/filtersList.dart';
 
 class FilterPremiumView extends StatefulWidget {
   @override
